Added comments to readme and config files
Readme
Uses lots of dynamic SQL, but all of it within sqlite3 -- this is a
sentence fragment.
, but only adds tables, doesn't modify existing ones with the exception
of empty tables. -- Statements are somewhat contradictory
--concat which concatenates the code variables -- Is this where you put
all of them in the same cell in a json like format? Should be further
explained what you mean by 'concatenate'.
* **rulesdef.csv -- this file is no longer necessary because of the
Config file:
adding comments inline-- this file should NOT be considered deployable.
It is marked up with comments identified per request with <!-- -->